requirements.txt
aiohttp==3.5.4
appnope==0.1.0
appscript==1.1.0
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
avro==1.9.0
avro-python3==1.9.0
backcall==0.1.0
bayesian-optimization==1.0.1
beautifulsoup4==4.7.1
bleach==3.0.2
boto==2.49.0
boto3==1.9.110
botocore==1.12.110
bz2file==0.98
category-encoders==2.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
colorama==0.4.0
confluent-kafka==0.11.6
cycler==0.10.0
cymem==2.0.2
Cython==0.29
cytoolz==0.9.0.1
dask==1.2.0
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.9
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
explicit==0.1.3
fastavro==0.21.21
Flask==1.1.1
funcy==1.11
future==0.17.1
gensim==3.7.1
h5py==2.9.0
idna==2.8
implicit==0.3.8
instaloader==4.2.4
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.12.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kafka-python==1.3.5
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyboard==0.13.3
kiwisolver==1.0.1
ksql==0.5.1.1
lightfm==1.15
lxml==4.3.3
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.4
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==4.5.2
murmurhash==1.0.2
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.4
nose==1.3.7
notebook==5.7.0
numexpr==2.6.9
numpy==1.16.2
oauthlib==3.0.1
openpyxl==2.6.2
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pbr==5.2.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pipreqs==0.4.9
plac==0.9.6
pluggy==0.8.0
preshed==2.0.1
prettytable==0.7.2
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.1.0
ptvsd==4.2.4
ptyprocess==0.6.0
py==1.7.0
pyaml==18.11.0
pygame==1.9.4
pygame-menu==1.95.2
Pygments==2.2.0
pyLDAvis==2.1.2
pyltr==0.2.4
pynput==1.4.2
pyobjc==5.2
pyobjc-core==5.2
pyobjc-framework-Accounts==5.2
pyobjc-framework-AddressBook==5.2
pyobjc-framework-AdSupport==5.2
pyobjc-framework-AppleScriptKit==5.2
pyobjc-framework-AppleScriptObjC==5.2
pyobjc-framework-ApplicationServices==5.2
pyobjc-framework-Automator==5.2
pyobjc-framework-AVFoundation==5.2
pyobjc-framework-AVKit==5.2
pyobjc-framework-BusinessChat==5.2
pyobjc-framework-CalendarStore==5.2
pyobjc-framework-CFNetwork==5.2
pyobjc-framework-CloudKit==5.2
pyobjc-framework-Cocoa==5.2
pyobjc-framework-Collaboration==5.2
pyobjc-framework-ColorSync==5.2
pyobjc-framework-Contacts==5.2
pyobjc-framework-ContactsUI==5.2
pyobjc-framework-CoreAudio==5.2
pyobjc-framework-CoreAudioKit==5.2
pyobjc-framework-CoreBluetooth==5.2
pyobjc-framework-CoreData==5.2
pyobjc-framework-CoreLocation==5.2
pyobjc-framework-CoreMedia==5.2
pyobjc-framework-CoreMediaIO==5.2
pyobjc-framework-CoreML==5.2
pyobjc-framework-CoreServices==5.2
pyobjc-framework-CoreSpotlight==5.2
pyobjc-framework-CoreText==5.2
pyobjc-framework-CoreWLAN==5.2
pyobjc-framework-CryptoTokenKit==5.2
pyobjc-framework-DictionaryServices==5.2
pyobjc-framework-DiscRecording==5.2
pyobjc-framework-DiscRecordingUI==5.2
pyobjc-framework-DiskArbitration==5.2
pyobjc-framework-DVDPlayback==5.2
pyobjc-framework-EventKit==5.2
pyobjc-framework-ExceptionHandling==5.2
pyobjc-framework-ExternalAccessory==5.2
pyobjc-framework-FinderSync==5.2
pyobjc-framework-FSEvents==5.2
pyobjc-framework-GameCenter==5.2
pyobjc-framework-GameController==5.2
pyobjc-framework-GameKit==5.2
pyobjc-framework-GameplayKit==5.2
pyobjc-framework-ImageCaptureCore==5.2
pyobjc-framework-IMServicePlugIn==5.2
pyobjc-framework-InputMethodKit==5.2
pyobjc-framework-InstallerPlugins==5.2
pyobjc-framework-InstantMessage==5.2
pyobjc-framework-Intents==5.2
pyobjc-framework-IOSurface==5.2
pyobjc-framework-iTunesLibrary==5.2
pyobjc-framework-LatentSemanticMapping==5.2
pyobjc-framework-LaunchServices==5.2
pyobjc-framework-libdispatch==5.2
pyobjc-framework-LocalAuthentication==5.2
pyobjc-framework-MapKit==5.2
pyobjc-framework-MediaAccessibility==5.2
pyobjc-framework-MediaLibrary==5.2
pyobjc-framework-MediaPlayer==5.2
pyobjc-framework-MediaToolbox==5.2
pyobjc-framework-ModelIO==5.2
pyobjc-framework-MultipeerConnectivity==5.2
pyobjc-framework-NaturalLanguage==5.2
pyobjc-framework-NetFS==5.2
pyobjc-framework-Network==5.2
pyobjc-framework-NetworkExtension==5.2
pyobjc-framework-NotificationCenter==5.2
pyobjc-framework-OpenDirectory==5.2
pyobjc-framework-OSAKit==5.2
pyobjc-framework-Photos==5.2
pyobjc-framework-PhotosUI==5.2
pyobjc-framework-PreferencePanes==5.2
pyobjc-framework-PubSub==5.2
pyobjc-framework-QTKit==5.2
pyobjc-framework-Quartz==5.2
pyobjc-framework-SafariServices==5.2
pyobjc-framework-SceneKit==5.2
pyobjc-framework-ScreenSaver==5.2
pyobjc-framework-ScriptingBridge==5.2
pyobjc-framework-SearchKit==5.2
pyobjc-framework-Security==5.2
pyobjc-framework-SecurityFoundation==5.2
pyobjc-framework-SecurityInterface==5.2
pyobjc-framework-ServiceManagement==5.2
pyobjc-framework-Social==5.2
pyobjc-framework-SpriteKit==5.2
pyobjc-framework-StoreKit==5.2
pyobjc-framework-SyncServices==5.2
pyobjc-framework-SystemConfiguration==5.2
pyobjc-framework-UserNotifications==5.2
pyobjc-framework-VideoSubscriberAccount==5.2
pyobjc-framework-VideoToolbox==5.2
pyobjc-framework-Vision==5.2
pyobjc-framework-WebKit==5.2
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.0.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
regex==2018.1.10
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
scikit-learn==0.20.0
scikit-surprise==1.0.6
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==1.8.0
soupsieve==1.8
spacy==2.0.18
statsmodels==0.9.0
stop-words==2018.7.23
surprise==0.1
telepot==12.7
terminado==0.8.1
testpath==0.4.2
thinc==6.12.1
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
tweepy==3.7.0
twython==3.7.0
ujson==1.35
urllib3==1.24.1
virtualenv==16.5.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.4.2
wrapt==1.10.11
xgboost==0.81
xlrd==1.2.0
yarg==0.1.9
yarl==1.3.0
